requirements/docs.txt
sphinx
sphinx-gallery
sphinx-autoapi
numpydoc>=1.1.0
